requirements.txt
chumpy
comet-ml==3.40.0
cython==0.29.20
easydict
ipdb
kornia==0.6.12
loguru
matplotlib
numpy
open3d
opencv-contrib-python-headless
opencv-python
pillow
pygit2==1.10.1
pymeshlab==2022.2.post4
pyparsing
pytorch-lightning==1.5.7
scipy
scikit-learn
tqdm
trimesh
